Better storage of WS messages
Instead of all messages being stored in a single redis sorted set, they
are stored by repository+topic.
So there is a lot less work to do on reconciliation to know what
messages to retrieve.

Also, the `ws` library was totally rewritten and is fully tested (it's
the first part of gim that is tested !)