Manage our signals after saving m2m
Because logically, m2m are saved after the object, because this object
is needed.
But the signals are triggered when saving the object, ie before the m2m,
so our siangl handlers have incomplete data.
Now, in issue view and github fetching, we have a new status "saving"
when saving before the m2m, and this status is set to "fetched" after
that.