Correct time-ago pbm with cached date
It occurs when the page is loaded from the cache with an old date attribute in the body. 
Now, the current server date is sent on ping, and the dates of the time_ago module are updated. So on the next date replacement, the new base dates are used, and everything is normal. It could take up to one minute, though.