Issue views does not depend on a repository anymore
This also means that there can only be rendered by ajax. If not, a redirect will be done to the repository view or subscription view.
Note that IssueSummaryView and IssueWithoutDetailsView have a part commented out that depended on the repository view: the check if the issue matched the current repository filter. More refactor will be needed for this.