Stop displaying twice an issue when updating project
We receive two update events: from the card and the issue.
Both are called at the same time, before one marking the update to be
currently happening.
So both trigger on_issue_create_alert and the issue is displayed twice.
Now in the callback to the get we check that the issue is not already
here.