Allow to run crossbar in different port on dev
The connection to redis/limpyd is shared and on crossbar, connections
are run in the same process, so they share the same connection.
But if one request is currently doing some pipeline stuff while another
request comes, this second one will be in pipeline mode too.

It doesn't happen in production because gunicorn manages the threads
differently (I guess).

But in dev mode, we should have crossbar managing only web-socket stuff,
and use the django runserver with `--nothreading` for the frontend until
a solution is found.

cf issue on limpyd https://github.com/limpyd/redis-limpyd/issues/97