Cache label-type for label using name, not typed name
The orignal name of the label, given as argument, was reset when calling
`get_name_and_order`.

We encountered the problem really late, with ansible, that had a
"module" label, without type, then later added many labels starting with
"new_", including "new_module".

So for the "new_module" label, the typed_name "module" was used for the
cache, so used also for module. Or the reverse, no type for "new_module"
if "module" was used first.

Because of this, a lot of issues cached template were regenerated all
the time.