Add code to manage a user with an expired token
If the token is expired, it will be removed.
If it's the default token for the user (in db), we'll find another and
use it.
If no other token, the user has no token anymore.
Then all jobs depending on this user (self jobs, or jobs where he is the
only one with the correct permission) are canceled